- started working on win32 compatibility - this should have no effect on the posix version
- #ifdef-ed some #includes
- replaced bzero with memset
- replaced a bunch of ints with size_t to avoid warnings
- all the open() calls now take 3 parameters, never 2 (this should have no effect on the posix version)
- moved some variables to the start of the block
- don't have any more declare + initialize on one line


git-svn-id: svn://littlesvr.ca/prog/isomaster/src/bk@457 ea4051ab-32fe-0310-aa8c-8a8c8f7b8cea

+++ bkCache.c
@@ -24,7 +24,6 @@
 #include <sys/timeb.h>
 #include <sys/types.h>
 #include <sys/stat.h>
-#include <unistd.h>
 
 #include "bkInternal.h"
 #include "bkCache.h"
@@ -48,7 +47,7 @@ off_t wcSeekTell(VolInfo* volInfo)
     return lseek(volInfo->imageForWriting, 0, SEEK_CUR);
 }
 
-int wcWrite(VolInfo* volInfo, const char* block, off_t numBytes)
+int wcWrite(VolInfo* volInfo, const char* block, size_t numBytes)
 {
     int rc;
     rc = write(volInfo->imageForWriting, block, numBytes);
